pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs>=0.4",
  "hatchling>=1.25",
]

[project]
name = "hope-api-auth"
description = "Provides API Authentication and Authorization within HOPE Ecosystem"
readme = "README.md"
license = "Apache-2.0 OR MIT"
authors = [
  { name = "UNICEF", email = "rapidpro@unicef.org" },
]
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [
  "version",
]
dependencies = [
  "django",
  "django-admin-extra-buttons",
  "django-adminfilters",
  "django-smart-admin",
  "djangorestframework",
  "psycopg2-binary",
  "swapper",
]

urls.Homepage = "https://github.com/unicef/hope-api-auth"

[dependency-groups]
dev = [
  "django-regex",
  "django-webtest",
  "factory-boy",
  "mock",
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-django",
  "pytest-echo",
  "pytest-factoryboy",
  "tox",
]

[tool.hatch]
build.hooks.vcs.version-file = "src/hope_api_auth/version.py"
version.source = "vcs"
raw-options = { root = "../.." }

[tool.hatch.build.targets.sdist]
ignore-vcs = true
include = [
  "src/hope_api_auth/**/*.*",
]

[tool.hatch.build.targets.wheel]
packages = [ "src/hope_api_auth" ]

[tool.uv]
package = true

[tool.nitpick]
style = [
  "github://unicef/hope-code-conventions@main/django/django.toml",
]
cache = "1 day"
